[SASU-0038] Code simplified
Hardcoded strings removed & one line function removed for simplicity
