+++ SampleAppSwiftUI/Network/WebSocket/RequestModels/SubscriptionRequest.swift
@@ -1,0 +1,13 @@
+//
+//  WebSocketSubscriptionRequest.swift
+//  SampleAppSwiftUI
+//
+//  Created by Uslu, Teyhan on 19.04.2023.
+//
+
+import Foundation
+
+struct SubscriptionRequest: Codable {
+    let action: String
+    let subs: [String]
+}
