+++ SampleAppSwiftUI/Network/WebSocket/RequestModels/SubscriptionRequestAction.swift
@@ -1,0 +1,13 @@
+//
+//  SubscriptionRequestAction.swift
+//  SampleAppSwiftUI
+//
+//  Created by Sucu, Ege on 18.05.2023.
+//
+
+import Foundation
+
+enum SubscriptionRequestAction: String {
+    case add = "SubAdd"
+    case remove = "SubRemove"
+}
