+++ SampleAppSwiftUI/Resources/Constants/Strings+Generated.swift
@@ -10,6 +10,10 @@ import Foundation
 // swiftlint:disable explicit_type_interface function_parameter_count identifier_name line_length
 // swiftlint:disable nesting type_body_length type_name vertical_whitespace_opening_braces
 public enum Strings {
+  /// 5~CCCAGG~%@~%@
+  public static func coinPreRequest(_ p1: Any, _ p2: Any) -> String {
+    return Strings.tr("Localizable", "CoinPreRequest", String(describing: p1), String(describing: p2), fallback: "5~CCCAGG~%@~%@")
+  }
   /// Favorites
   public static let favorites = Strings.tr("Localizable", "Favorites", fallback: "Favorites")
   /// Localizable.strings
