+++ SampleAppSwiftUI/Resources/Mocks/Sample-Subscription.geojson
@@ -1,0 +1,4 @@
+{
+   "action": "SubAdd",
+   "subs": ["5~CCCAGG~BTC~USD"]
+}
