+++ SampleAppSwiftUI/Resources/de.lproj/Localizable.strings
@@ -8,3 +8,4 @@
 
 "Hello, World!" = "Hallo";
 "Favorites" = "Favoriten";
+"CoinPreRequest" = "5~CCCAGG~%@~%@";
