+++ SampleAppSwiftUI/Resources/en.lproj/Localizable.strings
@@ -8,3 +8,4 @@
 
 "Hello, World!" = "Hello";
 "Favorites" = "Favorites";
+"CoinPreRequest" = "5~CCCAGG~%@~%@";
