+++ SampleAppSwiftUI/Resources/tr.lproj/Localizable.strings
@@ -8,3 +8,4 @@
 
 "Hello, World!" = "Merhaba";
 "Favorites" = "Favoriler";
+"CoinPreRequest" = "5~CCCAGG~%@~%@";
